+++ mobileapp/adapters/lib/presenters.dart
@@ -120,6 +120,7 @@ class ApplicationWidePresenter implements PresentationBoundary {
     SummitDetailsModel model = SummitDetailsModel(
       selectedSummit.id,
       selectedSummit.name,
+      selectedSummit.sector,
       canShowOnMap: selectedSummit.position != null,
     );
     ui.showSummitDetails(model);
@@ -177,6 +178,10 @@ class ApplicationWidePresenter implements PresentationBoundary {
       selectedRoute.id,
       selectedRoute.routeName,
       _labelCreator.createGradeLabel(selectedRoute),
+      <ListViewItem>[
+        for (final Directions directions in selectedRoute.directions)
+          ListViewItem(directions.content, bottomLine: '(Quelle: ${directions.source})'),
+      ],
     );
     ui.showRouteDetails(model);
   }